tests/Elksoft.Converters.Tests/Elksoft.Converters.ImplicitNumericConverters/ImplicitByteToCharConverterTests.cs
using System;
using FluentAssertions;
using Moq;
using Xunit;

namespace Elksoft.Converters.ImplicitNumericConverters
{
    public class ImplicitByteToCharConverterTests
    {
        [Fact]
        public static void AcceptsNull_IsFalse()
        {
            var converter = new ImplicitByteToCharConverter();
            converter.AcceptsNull.Should().BeFalse();
        }

        [Fact]
        public static void IsExplicit_IsFalse()
        {
            var converter = new ImplicitByteToCharConverter();
            converter.IsExplicit.Should().BeFalse();
        }

        public static TheoryData<Byte, Char> Convert_ReturnsExpected_Data()
        {
            return new TheoryData<Byte, Char>() {
                { 0, '\u0000' },
                { 1, '\u0001' },
                { 127, (Char)SByte.MaxValue },
                { Byte.MaxValue, (Char)Byte.MaxValue },
            };
        }

        [Theory]
        [MemberData(nameof(Convert_ReturnsExpected_Data))]
        public static void Convert_ReturnsExpected(Byte input, Char output)
        {
            var mockProvider = new Mock<IFormatProvider>(MockBehavior.Strict);

            var converter = new ImplicitByteToCharConverter();
            converter.Convert(input, null).Should().Be(output);
            converter.Convert(input, mockProvider.Object).Should().Be(output);

            mockProvider.VerifyAll();
        }
    }
}
